+++ src/components/SidePanel/SwitchPoints.jsx
@@ -35,14 +35,16 @@
  */
 
 import React from 'react';
-import { string } from 'prop-types';
 import { useDispatch, useSelector } from 'react-redux';
 import Button from 'react-bootstrap/Button';
 import Row from 'react-bootstrap/Row';
 import Col from 'react-bootstrap/Col';
 
-import { Slider, Toggle } from 'pc-nrfconnect-shared';
-import Collapse from './Collapse';
+import {
+    CollapsibleGroup,
+    Slider,
+    Toggle,
+} from '../../from_pc-nrfconnect-shared';
 
 import {
     spikeFilteringToggle,
@@ -57,7 +59,7 @@ import {
 } from '../../reducers/switchingPointsReducer';
 import { appState } from '../../reducers/appReducer';
 
-const SwitchPoints = ({ eventKey }) => {
+const SwitchPoints = () => {
     const dispatch = useDispatch();
 
     const {
@@ -76,7 +78,7 @@ const SwitchPoints = ({ eventKey }) => {
     }
 
     return (
-        <Collapse heading="SWITCH LEVELS" eventKey={eventKey}>
+        <CollapsibleGroup heading="Switch levels">
             {capabilities.ppkSwitchPointUp && (
                 <>
                     <span title="Set dynamic range switching levels. See user guide for details.">
@@ -132,12 +134,8 @@ const SwitchPoints = ({ eventKey }) => {
                     variant="secondary"
                 />
             )}
-        </Collapse>
+        </CollapsibleGroup>
     );
-};
-
-SwitchPoints.propTypes = {
-    eventKey: string.isRequired,
 };
 
 export default SwitchPoints;