+++ src/reducers/appReducer.js
@@ -59,7 +59,7 @@ const TOGGLE_ADVANCED_MODE = 'TOGGLE_ADVANCED_MODE';
 const TOGGLE_SAVE_CHOICE_DIALOG = 'TOGGLE_SAVE_CHOICE_DIALOG';
 const SHOW_EXPORT_DIALOG = 'SHOW_EXPORT_DIALOG';
 const HIDE_EXPORT_DIALOG = 'HIDE_EXPORT_DIALOG';
-const TOGGLE_SAVE = 'TOGGLE_SAVE';
+const TOGGLE_SAVE_FUNCTIONALITY = 'TOGGLE_SAVE_FUNCTIONALITY';
 
 export const toggleAdvancedModeAction = () => ({ type: TOGGLE_ADVANCED_MODE });
 export const samplingStartAction = () => ({ type: SAMPLING_STARTED });
@@ -90,6 +90,10 @@ export const showExportDialog = () => ({ type: SHOW_EXPORT_DIALOG });
 export const hideExportDialog = () => ({ type: HIDE_EXPORT_DIALOG });
 export const toggleSaveAction = () => ({ type: TOGGLE_SAVE });
 
+export const toggleSaveFunctionality = () => ({
+    type: TOGGLE_SAVE_FUNCTIONALITY,
+});
+
 export default (state = initialState, { type, ...action }) => {
     switch (type) {
         case DEVICE_OPENED: {
@@ -110,6 +114,8 @@ export default (state = initialState, { type, ...action }) => {
             return { ...state, rttRunning: true };
         case TOGGLE_ADVANCED_MODE:
             return { ...state, advancedMode: !state.advancedMode };
+        case TOGGLE_SAVE_FUNCTIONALITY:
+            return { ...state, enableSave: !state.enableSave };
         case TOGGLE_SAVE_CHOICE_DIALOG:
             return {
                 ...state,
@@ -134,3 +140,9 @@ export default (state = initialState, { type, ...action }) => {
 };
 
 export const appState = ({ app }) => app.app;
+
+export const advancedMode = state => state.app.app.advancedMode;
+export const deviceOpen = state =>
+    Object.keys(state.app.app.capabilities).length > 0;
+
+export const currentPane = state => state.appLayout.currentPane;