+++ Changelog.md
@@ -1,3 +1,186 @@
-## Version 3.0.0
-### Changed
-- Complete rewrite of the application. This version supports the old power profiler kit, in addition to the newly released PPK2 hardware.
+## Unreleased
+### Updates
+- Add screenshot functionality
+
+## Version 3.0.0-beta.13
+### Updates
+- Also remove loading while saving is removed.
+
+## Version 3.0.0-beta.12
+### Updates
+- When the save/export button is disabled explain in tooltip why
+- Unified the naming of the last range in the spike filter to be range
+  number 5. If you changed the coefficient for range 5, you will have
+  to set it again, as the old setting is not restored.
+- Add 3 second button
+- Temporarily disable saving because it breaks compatibility. Export remains.
+
+## Version 3.0.0-beta.11
+### Updates
+- Set max buffer to 5min
+- Update digital channel visibility to 10s window limit
+
+## Version 3.0.0-beta.10
+### Updates
+- Log spike filter only in advanced mode
+- Replaced 1ms window with 10s window
+- Add more tooltips
+
+## Version 3.0.0-beta.9
+### Updates
+- Updated firmware to 1.0.0-1b5c1ac
+- Added tooltips
+- Hid the voltage slider if voltage is not settable
+- Updated label of trigger button
+
+## Version 3.0.0-beta.8
+### Updates
+- Fixed trigger range
+- Added animation to running sampling and trigger state
+- Reintroduced advanced mode
+
+## Version 3.0.0-beta.7
+### Updates
+- Update DUT power toggle label
+- Prevent data loss from child process
+- Fix window title update
+
+## Version 3.0.0-beta.6
+### Updates
+- Updated firmware to d38cae1
+- Disabled power output by default
+
+## Version 3.0.0-beta.5
+### Updates
+- Updated firmware to cdbd883 and filter defaults
+
+## Version 3.0.0-beta.4
+### Updates
+- Removed trigger level slider and updated the input line behaviour
+- Fixed displaying NaN values
+- Fixed updating trigger level while trigger is active
+- Log trigger level change from chart
+- Fixed currentVdd for calculations
+- Updated window title with filename or serialnumber
+
+## Version 3.0.0-beta.3
+### Updates
+- Fix broken build
+
+## Version 3.0.0-beta.2
+### Updates
+- Updated for new rules of shared@4.9.3
+- Added supporting component for bootstrap tooltips
+- Updated ppk2 firmware to 7d45fe4
+
+## Version 3.0.0-beta.1
+### Updates
+- Sharp crosshair, trigger, marker lines
+- Updated start button label for PPK1
+- Updated padding, font-size of calculated values
+
+## Version 3.0.0-alpha.13
+### Updates
+- Fixed selected area in the chart
+- Fixed crosshair alignment
+- Pan keeps the window length
+- Better y-range limit
+
+## Version 3.0.0-alpha.12
+### Updates
+- New trigger handle
+- Limit y-axis range
+- Changed button label capitalization
+- Add log when single trigger received
+
+## Version 3.0.0-alpha.11
+### Updates
+- Added shift+touchpad events for better zooming
+- Fix timespan height and alignment
+- Fix collapsible icon alignment
+- Changed omega sign symbol
+- Added info about digital channels to see
+- Use svg handles on selection
+
+## Version 3.0.0-alpha.10
+### Updates
+- Disable trigger level for external trigger
+- Update stat-box heading style
+- Use only one color for trigger level
+- Update sidepanel initial layout
+- Responsive width of chart top buttons
+- Update trigger, crosshair and markers in chart
+
+## Version 3.0.0-alpha.9
+### Updates
+- Updated firmware
+- Added defaults button to spike filter
+- Moved collapsible icon to right
+- Update control styles in sidepanel
+- Add dot indicator to power mode selector
+- Enhanced chart elements styles
+
+## Version 3.0.0-alpha.8
+### Updates
+- Minor style updates
+- Added progress bar to export dialog, made export cancellable
+- Reset y-range when turning y-lock off
+
+## Version 3.0.0-alpha.7
+### Updates
+- Updated point radius in main chart
+- Fixed y-axis lock when limits are unset
+- Rearranged sidepanel
+- Updated style of trigger level handle
+
+## Version 3.0.0-alpha.6
+### Updates
+- Fixed trigger level rounding
+- Enabled crosshair only when chart is paused
+- Added support for panning with touchpad
+- Refactored trigger control
+
+## Version 3.0.0-alpha.5
+### Updates
+- Added lock y-axis toggle
+- Removed hidden advanced mode
+- Introduced custom trigger range visual
+- Added snapping to crosshair, removed tooltips
+
+## Version 3.0.0-alpha.4
+### Updates
+- Moved save/export to top, fixed collapse defaults
+- Toggle live mode when trigger starts
+- Removed digital options for v1 export
+- Added selection handle lines, limited area
+- Updated crosshair color
+
+## Version 3.0.0-alpha.3
+### Updates
+- Chart live view toggle always enabled
+- Moved digital channel toggles to display options
+- Rearraged trigger controls
+- Renamed trigger window to trigger length
+- Fixed updating trigger controls
+- Only send trigger level before trigging started
+- Independent collapsibles in the side panel
+
+## Version 3.0.0-alpha.2
+### Updates
+- Save last directory used
+
+## Version 3.0.0-alpha.1
+### Updates
+- First automated test release
+
+## Version 2.3.1
+### Bugfixes
+- Fix updating voltage regulator #105
+
+## Version 2.3.0
+### Updates
+- Updated RTT API calls for nRFConnect 3.3 #103
+
+## Version 2.2.0
+### Updates
+- Updated to React Bootstrap 4